LayoutTests/platform/chromium-mac/fast/dom/HTMLSelectElement/named-options-expected.txt

This tests that options elements are accessible by name from both a select element and the options collection, per HTML5.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Confirm that the option named 'test' is accessible from the select element
FAIL select1.test.toString() should be [object HTMLOptionElement]. Threw exception TypeError: Cannot call method 'toString' of undefined
FAIL select1.test.value should be Value. Threw exception TypeError: Cannot read property 'value' of undefined
Confirm that the option named 'test' is accessible from the options collection
PASS select1.options.test.toString() is "[object HTMLOptionElement]"
PASS select1.options.test.value is "Value"
Confirm that both options named 'test' are accessible from the select element
FAIL select2.test.length should be 2. Threw exception TypeError: Cannot read property 'length' of undefined
FAIL select2.test.toString() should be [object NodeList]. Threw exception TypeError: Cannot call method 'toString' of undefined
FAIL select2.test[0].value should be Value1. Threw exception TypeError: Cannot read property '0' of undefined
FAIL select2.test[1].value should be Value2. Threw exception TypeError: Cannot read property '1' of undefined
Confirm that both options named 'test' are accessible from the options collection
PASS select2.options.test.length is 2
PASS select2.options.test.toString() is "[object NodeList]"
PASS select2.options.test[0].value is "Value1"
PASS select2.options.test[1].value is "Value2"
PASS successfullyParsed is true

TEST COMPLETE
